projects/seattle-quantiphi-staging-dev-test/seattle-quantiphi-staging-dev-test/TMA/flex_tma/cloud_function/requirements.txt
# Function dependencies, for example:
# package>=version
google-api-python-client==1.12.8
google-cloud-storage==1.36.1
oauth2client==4.1.3